:sparkles: BST implemented for names, runtime ~0.2 seconds
Used insert-only binary search tree that adds name to dupes array if it already exists, otherwise adds it to the BST
