+++ names/names.py
@@ -1,5 +1,4 @@
 import time
-from binary_search_tree import BinarySearchTree
 
 start_time = time.time()
 
@@ -12,10 +11,10 @@ names_2 = f.read().split("\n")  # List containing 10000 names
 f.close()
 
 duplicates = []
-for name_1 in names_1:
-    for name_2 in names_2:
-        if name_1 == name_2:
-            duplicates.append(name_1)
+# for name_1 in names_1:
+#     for name_2 in names_2:
+#         if name_1 == name_2:
+#             duplicates.append(name_1)
 
 # BST pseudocode:
 """
@@ -24,6 +23,35 @@ for n in names_1+names_2: # O(n) runtime
     BUT insert appends it to the duplicate array instead if strictly equal (dupe)
     so BST.insert for the whole concatenated array would approach O(log n)
 """
+
+
+class BinarySearchTree:
+    def __init__(self, name):
+        self.name = name
+        self.left = None
+        self.right = None
+
+    def insert(self, name):
+        # greater names alphabetically become right child
+        if self.name.lower() < name.lower():
+            if self.right:
+                self.right.insert(name)
+            else:
+                self.right = BinarySearchTree(name)
+        # alphabetically lower names become left child
+        elif self.name.lower() > name.lower():
+            if self.left:
+                self.left.insert(name)
+            else:
+                self.left = BinarySearchTree(name)
+        else:  # names are equal--a duplicate
+            duplicates.append(name)
+
+
+bst = BinarySearchTree(names_1[0])
+
+for name in names_1[1:]+names_2:
+    bst.insert(name)
 
 end_time = time.time()
 print(f"{len(duplicates)} duplicates:\n\n{', '.join(duplicates)}\n\n")
